Simply trace reader interface by not throwing IOException
This removes the checked exception from the interface, which is never
needed by the implementations. The synthetic ones perform no I/O, while
the file-based ones utilize an abstract base class templates. Instead
that exception can be rethrown as unchecked, since it will halt the
simulator as non-recoverable.

This minor nit was because PMD now errors on simple aggregation loop,
which used a Stream.empty() as the base to concatinate on. Since that
variable was reassigned and not closed, its updated check would need
to be suppressed. Now that the checked exception is gone, it can be
replaced using flatMap to concat trace events.